+++ app/Console/Kernel.php
@@ -25,6 +25,7 @@ class Kernel extends ConsoleKernel
         \App\Console\Commands\MigrateTags::class,
         \App\Console\Commands\ImportCsv::class,
         \App\Console\Commands\ItemsExtractColors::class,
+        \App\Console\Commands\MatchAuthorities::class,
     ];
 
     /**