+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ### Added
+- authority matcher
+
+### Changed
+- use authority matcher in item detail
 - SGP into list of galleries in the info page
 
 ## [2.13.0] - 2020-11-27